+++ src/main/java/chess/ConsoleApplication.java
@@ -2,13 +2,13 @@ package chess;
 
 import chess.controller.console.ConsoleController;
 import chess.mysql.ChessGameManagerRepositoryImpl;
-import chess.mysql.dao.MysqlChessDao;
-import chess.mysql.dao.StaticMemoryChessDao;
+import chess.mysql.MysqlChessDao;
+import chess.mysql.StaticMemoryChessDao;
 import chess.service.ChessServiceImpl;
 
 import java.util.Objects;
 
-import static chess.mysql.dao.ChessConnection.getConnection;
+import static chess.mysql.ChessConnection.getConnection;
 
 public class ConsoleApplication {
     public static void main(String[] args) {
