+++ src/main/java/chess/SparkUIChessApplication.java
@@ -2,13 +2,13 @@ package chess;
 
 import chess.controller.web.SparkController;
 import chess.mysql.ChessGameManagerRepositoryImpl;
-import chess.mysql.dao.MysqlChessDao;
-import chess.mysql.dao.StaticMemoryChessDao;
+import chess.mysql.MysqlChessDao;
+import chess.mysql.StaticMemoryChessDao;
 import chess.service.ChessServiceImpl;
 
 import java.util.Objects;
 
-import static chess.mysql.dao.ChessConnection.getConnection;
+import static chess.mysql.ChessConnection.getConnection;
 import static spark.Spark.port;
 import static spark.Spark.staticFileLocation;
