+++ src/main/java/chess/chessgame/domain/room/game/ChessGameManager.java
@@ -4,7 +4,7 @@ import chess.chessgame.domain.room.game.board.Board;
 import chess.chessgame.domain.room.game.board.piece.attribute.Color;
 import chess.chessgame.domain.room.game.board.position.Position;
 import chess.chessgame.domain.room.game.statistics.ChessGameStatistics;
-import chess.controller.web.dto.PieceDto;
+import chess.controller.web.PieceDto;
 
 import java.util.Map;
