+++ src/main/java/chess/chessgame/domain/room/game/RunningGameManager.java
@@ -5,7 +5,7 @@ import chess.chessgame.domain.room.game.board.piece.attribute.Color;
 import chess.chessgame.domain.room.game.board.position.Position;
 import chess.chessgame.domain.room.game.statistics.ChessGameStatistics;
 import chess.chessgame.domain.room.game.statistics.MatchResult;
-import chess.controller.web.dto.PieceDto;
+import chess.controller.web.PieceDto;
 
 import java.util.Map;
