+++ src/main/java/chess/controller/web/ExceptionAdvice.java
@@ -1,6 +1,5 @@
 package chess.controller.web;
 
-import chess.controller.web.dto.ErrorMessageResponseDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
