+++ src/main/java/chess/controller/web/RoomRequestDto.java
@@ -1,4 +1,4 @@
-package chess.controller.web.dto;
+package chess.controller.web;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
