+++ src/main/java/chess/mysql/UserDao.java
@@ -1,4 +1,4 @@
-package chess.mysql.dao;
+package chess.mysql;
 
 import chess.chessgame.domain.room.game.board.piece.attribute.Color;
 import chess.chessgame.domain.room.user.User;
