+++ src/test/java/chess/controller/web/GameRestControllerTest.java
@@ -1,6 +1,5 @@
 package chess.controller.web;
 
-import chess.controller.web.dto.RoomRequestDto;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
