+++ src/test/java/chess/mysql/dao/MysqlChessDaoTest.java
@@ -5,7 +5,9 @@ import chess.chessgame.domain.room.game.ChessGameManagerFactory;
 import chess.chessgame.domain.room.game.board.Square;
 import chess.chessgame.domain.room.game.board.piece.Pawn;
 import chess.chessgame.domain.room.game.board.position.Position;
-import chess.mysql.dao.dto.ChessGameDto;
+import chess.mysql.ChessConnection;
+import chess.mysql.ChessGameDto;
+import chess.mysql.MysqlChessDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
